Solid colour block initialisation for AFBC headers
From spec(AFBC 1.2):
If a complete superblock has only one color on all its pixels, it can be encoded using a header and no payload data. This storage method is permitted for superblock_layout 0, 3, 4, or 7 with 64 bits per
pixel or less.
When the data is encoded using a header with no payload data, the value of the pixel is stored least significant bit aligned in bits[127:64] of the header.
AFBC supports up to four solid color block components. These components are stored in the header that is tightly packed in order from least significant bit to most significant bit. The components are stored in
the following order, from least significant bit to most significant bit: RGBA. The rest bits are 0s.

Bug: 349318334
Change-Id: Ie97b16156ac55c3d1d2b639a09733ed5211a3fed